fix: keep manual implementations, apply only naming/type changes
- macros.rs cleared (consistent with ffi-implementation macro removal)
- handlers/payments.rs: FFIRequestData→FfiRequestData, common_crate→ucs_env,
  authorize_req→authorize_req_transformer, authorize_res→authorize_res_transformer,
  Option<MaskedMetadata> unwrap pattern, correct error return type on authorize_res_handler
- services/payments.rs: same renames + PaymentStatus::Pending→Failure in all error paths,
  no macro invocations
