+++ backend/ffi/src/handlers/payments.rs
@@ -1,26 +1,33 @@
 pub const EMBEDDED_DEVELOPMENT_CONFIG: &str = include_str!("../../../../config/development.toml");
 // pub mod napi_handler;
-use crate::macros::payment_flow_handler;
-use grpc_api_types::payments::{
-    PaymentServiceAuthorizeRequest, PaymentServiceAuthorizeResponse, PaymentServiceCaptureRequest,
-};
 
-use crate::services::payments::{
-    authorize_req_transformer, authorize_res_transformer, capture_req_transformer,
-};
+use grpc_api_types::payments::{PaymentServiceAuthorizeRequest, PaymentServiceAuthorizeResponse};
+
+use crate::services::payments::{authorize_req_transformer, authorize_res_transformer};
 
 use crate::types::FfiRequestData;
 use domain_types::payment_method_data::DefaultPCIHolder;
 
-// Generate authorize_req_handler using payment_flow_handler! macro
-payment_flow_handler!(
-    authorize_req_handler,
-    authorize_req_transformer,
-    PaymentServiceAuthorizeRequest,
-    DefaultPCIHolder
-);
+// authorize_req handler
+pub fn authorize_req_handler(
+    request: FfiRequestData<PaymentServiceAuthorizeRequest>,
+) -> Result<Option<common_utils::request::Request>, ucs_env::error::PaymentAuthorizationError> {
+    let metadata_payload = request.extracted_metadata;
+    let metadata_owned = request.masked_metadata.unwrap_or_default();
+    let metadata = &metadata_owned;
+    let payload = request.payload;
+    let config = crate::utils::load_config(EMBEDDED_DEVELOPMENT_CONFIG)?;
 
-// Generate authorize_res_flow handler
+    authorize_req_transformer::<DefaultPCIHolder>(
+        payload,
+        &config,
+        metadata_payload.connector,
+        metadata_payload.connector_auth_type,
+        metadata,
+    )
+}
+
+// authorize_res handler
 pub fn authorize_res_handler(
     request: FfiRequestData<PaymentServiceAuthorizeRequest>,
     response: domain_types::router_response_types::Response,
@@ -40,11 +47,3 @@ pub fn authorize_res_handler(
         response,
     )
 }
-
-// Generate capture_req_handler using payment_flow_handler! macro
-payment_flow_handler!(
-    capture_req_handler,
-    capture_req_transformer,
-    PaymentServiceCaptureRequest,
-    DefaultPCIHolder
-);
