+++ backend/ffi/src/services/payments.rs
@@ -1,26 +1,93 @@
-use crate::macros::payment_flow;
 use external_services;
-use grpc_api_types::payments::{
-    PaymentServiceAuthorizeRequest, PaymentServiceAuthorizeResponse, PaymentServiceCaptureRequest,
-};
+use grpc_api_types::payments::{PaymentServiceAuthorizeRequest, PaymentServiceAuthorizeResponse};
 use ucs_env::error::PaymentAuthorizationError;
 
 use domain_types::{
-    connector_flow::{Authorize, Capture},
-    connector_types::{PaymentFlowData, PaymentsAuthorizeData, PaymentsCaptureData},
+    connector_flow::Authorize,
+    connector_types::{PaymentFlowData, PaymentsAuthorizeData},
     router_data::ErrorResponse,
     router_data_v2::RouterDataV2,
     utils::ForeignTryFrom,
 };
 
-// Generate authorize_req_transformer function using the payment_flow! macro
-payment_flow!(
-    authorize_req_transformer,
-    Authorize,
-    PaymentServiceAuthorizeRequest,
-    PaymentsAuthorizeData<T>,
-    "PAYMENT_AUTHORIZE_ERROR"
-);
+pub fn authorize_req_transformer<
+    T: domain_types::payment_method_data::PaymentMethodDataTypes
+        + Default
+        + Eq
+        + std::fmt::Debug
+        + Send
+        + serde::Serialize
+        + serde::de::DeserializeOwned
+        + Clone
+        + Sync
+        + domain_types::types::CardConversionHelper<T>
+        + 'static,
+>(
+    payload: PaymentServiceAuthorizeRequest,
+    config: &std::sync::Arc<ucs_env::configs::Config>,
+    connector: domain_types::connector_types::ConnectorEnum,
+    connector_auth_details: domain_types::router_data::ConnectorAuthType,
+    metadata: &common_utils::metadata::MaskedMetadata,
+) -> Result<Option<common_utils::request::Request>, PaymentAuthorizationError> {
+    // connector integration trait
+    let connector_data: connector_integration::types::ConnectorData<T> =
+        connector_integration::types::ConnectorData::get_connector_by_name(&connector);
+
+    let connector_integration: interfaces::connector_integration_v2::BoxedConnectorIntegrationV2<
+        '_,
+        Authorize,
+        PaymentFlowData,
+        PaymentsAuthorizeData<T>,
+        domain_types::connector_types::PaymentsResponseData,
+    > = connector_data.connector.get_connector_integration_v2();
+
+    // Create PaymentFlowData from the payload
+    let payment_flow_data =
+        PaymentFlowData::foreign_try_from((payload.clone(), config.connectors.clone(), metadata))
+            .map_err(|err| {
+            tracing::error!(error = ?err, "Failed to create PaymentFlowData");
+            PaymentAuthorizationError::new(
+                grpc_api_types::payments::PaymentStatus::Failure,
+                Some(err.to_string()),
+                Some("PAYMENT_AUTHORIZE_ERROR".to_string()),
+                None,
+            )
+        })?;
+
+    // Create flow-specific request data
+    let payment_request_data: PaymentsAuthorizeData<T> =
+        PaymentsAuthorizeData::foreign_try_from(payload.clone()).map_err(|err| {
+            tracing::error!(error = ?err, "Failed to create payment request data");
+            PaymentAuthorizationError::new(
+                grpc_api_types::payments::PaymentStatus::Failure,
+                Some(err.to_string()),
+                Some("PAYMENT_AUTHORIZE_ERROR".to_string()),
+                None,
+            )
+        })?;
+
+    // Construct RouterDataV2 directly
+    let router_data = RouterDataV2 {
+        flow: std::marker::PhantomData,
+        resource_common_data: payment_flow_data,
+        connector_auth_type: connector_auth_details,
+        request: payment_request_data,
+        response: Err(ErrorResponse::default()),
+    };
+
+    // transform common request type to connector specific request type
+    let connector_request = connector_integration
+        .build_request_v2(&router_data.clone())
+        .map_err(|err| {
+            PaymentAuthorizationError::new(
+                grpc_api_types::payments::PaymentStatus::Failure,
+                Some(err.to_string()),
+                Some("PAYMENT_AUTHORIZE_ERROR".to_string()),
+                None,
+            )
+        })?;
+    Ok(connector_request)
+}
 
 pub fn authorize_res_transformer<
     T: domain_types::payment_method_data::PaymentMethodDataTypes
@@ -117,12 +184,3 @@ pub fn authorize_res_transformer<
         )
     })
 }
-
-// Generate capture_req_transformer function using the payment_flow! macro
-payment_flow!(
-    capture_req_transformer,
-    Capture,
-    PaymentServiceCaptureRequest,
-    PaymentsCaptureData,
-    "PAYMENT_CAPTURE_ERROR"
-);
